add: new database schema is being implemented and schema files separated
