+++ db/schema/accounts.ts
@@ -1,0 +1,33 @@
+import { integer, pgTable, text } from "drizzle-orm/pg-core";
+
+import { AdapterAccount } from "next-auth/adapters";
+import { users } from "./users";
+import { primaryKey, timestamp } from "drizzle-orm/pg-core";
+import { InferInsertModel, InferSelectModel } from "drizzle-orm";
+export const accounts = pgTable(
+  "account",
+  {
+    userId: text("userId")
+      .notNull()
+      .references(() => users.id, { onDelete: "cascade" }),
+    type: text("type").$type<AdapterAccount["type"]>().notNull(),
+    provider: text("provider").notNull(),
+    providerAccountId: text("providerAccountId").notNull(),
+    refresh_token: text("refresh_token"),
+    access_token: text("access_token"),
+    expires_at: integer("expires_at"),
+    token_type: text("token_type"),
+    scope: text("scope"),
+    id_token: text("id_token"),
+    session_state: text("session_state"),
+    createdAt: timestamp("createdAt", { mode: "date" }).defaultNow(),
+  },
+  (account) => ({
+    compoundKey: primaryKey({
+      columns: [account.provider, account.providerAccountId],
+    }),
+  })
+);
+
+export type SelectAccount = InferSelectModel<typeof accounts>;
+export type InsertAccount = InferInsertModel<typeof accounts>;
