+++ drizzle.config.ts
@@ -4,7 +4,7 @@ dotenv.config({
   path: ".env.local",
 });
 export default defineConfig({
-  schema: "./db/schema.ts",
+  schema: "./db/schema/*.ts",
   driver: "pg",
   dbCredentials: {
     connectionString: process.env.DATABASE_URL as string,
